+++ docs/file-structure.md
@@ -1,0 +1,5 @@
+# File Structure
+
+## Why do we use the `src/app/(app)` directory?
+
+Shipkit Bones is the foundation of Shipkit (a.k.a "Brains"), which uses PayloadCMS to manage content. We pre-emptively use the `(app)` directory to make installing and working with Payload more seamless.
